buildbot
this is a buildbot for mokee dev's and anyone that wants to build mokee.  it will build an infinate number of
devices in a row, as well as, if chosen by the user, sync repos before builds starts, run mka installclean,
make clean, crete an ota package, upload to mokeedev.com via ftp, and move the final  ROM (and OTA, if the user
choses to build one) and md5sum('s) to a storage dir.

all is automated, and no changes to the buildbot are needed, although there is a 'defaults' section at the top
that can be edited to speed things up.

if found to be usefull, a shortcut could be added to our base dir, so anyone that wants to build mokee would only
have to type 'buildbot' in, and as long as the dependencies for the device are up to date, everyinthing would then
be automated. for now use will be 'build/buildbot'

Change-Id: Ibe121941493200b6c3fce286f6b9a5625f2a5544