build: New build script
This is a build script for mokee dev's and anyone that wants to build mokee.  it will build an infinite number of
devices in a row, as well as, if chosen by the user, sync repos before builds starts, run mka installclean,
make clean, crete an ota package, upload to mokeedev.com via ftp, and move the final ROM (and OTA, if the user
choses to build one) and md5sum('s) to a storage dir.

All is automated, and no changes to the build script are needed, although there is a 'defaults' section at the top
that can be edited to speed things up (not recommended).

A shortcut was already added to our base dir, so anyone that wants to build mokee would only
have to type './mk' in, and as long as the dependencies for the device are up to date, everything would then
be automated.

Change-Id: I16957c75f7dabae4ff80322510795a50fbf2c29b