build: use the output directory actually in use.
It is not a good idea to output in the same directory as the source.
Someone may find it hard to manage or clean those files,
others may have two or more hard disks and want to reduce compile time.

We should move other output files(such as OTA and FTP files) to a separate directory in the future.

Change-Id: Ib31f94154938f9ada2d0e392b73db5ed61fb5c5b